requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-bootstrap-v5==1.0.11
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
oauthlib==3.2.2
Pillow == 9.3
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0